Use size_t instead of uint64_t for inlen.
Switched blake2b and blake2s to it, didn't touch
bp or sp versions, as they are unused for now, and
require more thinking for type conversion.

+++ impl/blake2.h
@@ -117,13 +117,13 @@ extern "C" {
   int blake2s_init( blake2s_state *S, const uint8_t outlen );
   int blake2s_init_key( blake2s_state *S, const uint8_t outlen, const void *key, const uint8_t keylen );
   int blake2s_init_param( blake2s_state *S, const blake2s_param *P );
-  int blake2s_update( blake2s_state *S, const uint8_t *in, uint64_t inlen );
+  int blake2s_update( blake2s_state *S, const uint8_t *in, size_t inlen );
   int blake2s_final( blake2s_state *S, uint8_t *out, uint8_t outlen );
 
   int blake2b_init( blake2b_state *S, const uint8_t outlen );
   int blake2b_init_key( blake2b_state *S, const uint8_t outlen, const void *key, const uint8_t keylen );
   int blake2b_init_param( blake2b_state *S, const blake2b_param *P );
-  int blake2b_update( blake2b_state *S, const uint8_t *in, uint64_t inlen );
+  int blake2b_update( blake2b_state *S, const uint8_t *in, size_t inlen );
   int blake2b_final( blake2b_state *S, uint8_t *out, uint8_t outlen );
 
   int blake2sp_init( blake2sp_state *S, const uint8_t outlen );
@@ -137,8 +137,8 @@ extern "C" {
   int blake2bp_final( blake2bp_state *S, uint8_t *out, uint8_t outlen );
 
   // Simple API
-  int blake2s( uint8_t *out, const void *in, const void *key, const uint8_t outlen, const uint64_t inlen, uint8_t keylen );
-  int blake2b( uint8_t *out, const void *in, const void *key, const uint8_t outlen, const uint64_t inlen, uint8_t keylen );
+  int blake2s( uint8_t *out, const void *in, const void *key, const uint8_t outlen, const size_t inlen, uint8_t keylen );
+  int blake2b( uint8_t *out, const void *in, const void *key, const uint8_t outlen, const size_t inlen, uint8_t keylen );
 
   int blake2sp( uint8_t *out, const void *in, const void *key, const uint8_t outlen, const uint64_t inlen, uint8_t keylen );
   int blake2bp( uint8_t *out, const void *in, const void *key, const uint8_t outlen, const uint64_t inlen, uint8_t keylen );
